cookbooks/financial-agent-demo/pyproject.toml
[project]
name = "financial-agent-demo"
version = "0.1.0"
description = "LangGraph workflow for financial research using Bigdata.com API"
authors = [
    {name = "Aakarsh Ramchandani", email = "support@bigdata.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bigdata-client>=2.17.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
